Added preliminary authenticatoin integration.
The code I have thus far successfully suppresses the login screen from
Activiti.

However, it is still a work in progress, in that IdentityService needs
to be integrated.

Another way to go about this is that we let Activiti maintains its own
user database, then we can simply provide a mapping between the Jenkins
user and Activiti user.